+++ tests/testthat/test-incidence.R
@@ -107,9 +107,16 @@ test_that("construction - Date input", {
   expect_message(x.i.trim  <- incidence(dat, first_date = 0),
                  "[0-9]+ observations outside of \\[0, [0-9]+\\] were removed."
   )
-  expect_message(x.d.trim  <- incidence(dat_dates, first_date = "2016-01-01"),
-                 "[0-9]+ observations outside of \\[2016-01-01, [-0-9]{10}\\] were removed."
-  )
+  expect_warning({
+  expect_message({
+    x.d.trim  <- incidence(dat_dates, first_date = "2016-01-01")
+  }, "[0-9]+ observations outside of \\[2016-01-01, [-0-9]{10}\\] were removed.")
+  }, "options\\(incidence.warn.first_date = FALSE\\)")
+  expect_message({
+  expect_failure(expect_warning({
+    x.d.trim  <- incidence(dat_dates, first_date = "2016-01-01")
+  }, "options\\(incidence.warn.first_date = FALSE\\)"))
+  }, "[0-9]+ observations outside of \\[2016-01-01, [-0-9]{10}\\] were removed.")
   x.7       <- incidence(dat_dates, 7L, standard = FALSE)
   x.7.iso   <- incidence(dat_dates, "week")
   x.7.week  <- incidence(dat_dates, "week", standard = FALSE)